requirements.txt
fastapi
uvicorn[standard]
pytesseract
Pillow
opencv-python-headless
pdf2image
requests
python-multipart
